# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: GCC-TBC Pro Typing Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /institutes:
    get:
      operationId: listInstitutes
      tags: [institutes]
      summary: List institutes
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Institutes list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstituteListResponse"

    post:
      operationId: createInstitute
      tags: [institutes]
      summary: Create a new institute
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InstituteInput"
      responses:
        "201":
          description: Created institute
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Institute"
  /institutes/{id}:
    get:
      operationId: getInstitute
      tags: [institutes]
      summary: Get institute by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Institute
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Institute"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateInstitute
      tags: [institutes]
      summary: Update institute
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InstituteUpdate"
      responses:
        "200":
          description: Updated institute
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Institute"

    delete:
      operationId: deleteInstitute
      tags: [institutes]
      summary: Delete institute
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /institutes/{id}/students:
    get:
      operationId: getInstituteStudents
      tags: [institutes]
      summary: Get students of an institute
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Students list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"

    post:
      operationId: addInstituteStudent
      tags: [institutes]
      summary: Institute admin directly creates a student account under their institute
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddInstituteStudentInput"
      responses:
        "201":
          description: Created student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /institutes/{id}/students/{userId}:
    delete:
      operationId: removeInstituteStudent
      tags: [institutes]
      summary: Remove a student from an institute — demotes to independent, never deletes the account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed (demoted to independent)
...
  /institutes/{id}/stats:
    get:
      operationId: getInstituteStats
      tags: [institutes]
      summary: Get institute analytics/stats
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Institute stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstituteStats"

  # ─── BATCHES ─────────────────────────────────────────────────────────────────
  /institutes/{id}/approve:
    post:
      operationId: approveInstitute
      tags: [institutes]
      summary: Owner approves an institute's account (inactive → active), e.g. after manual/offline payment verification
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Institute"
  /institutes/{id}/suspend:
    post:
      operationId: suspendInstitute
      tags: [institutes]
      summary: Owner suspends an institute's account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Suspended
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Institute"
  /institutes/{id}/premium:
    patch:
      operationId: setInstitutePremium
      tags: [institutes]
      summary: Owner manually grants or revokes Premium access for an institute (covers its whole roster) — bypasses payment entirely
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetPremiumInput"
      responses:
        "200":
          description: Updated institute
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Institute"
  /batches:
    get:
      operationId: listBatches
      tags: [institutes]
      summary: List batches for institute
      parameters:
        - name: instituteId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Batches list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Batch"

    post:
      operationId: createBatch
      tags: [institutes]
      summary: Create a new batch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchInput"
      responses:
        "201":
          description: Created batch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Batch"
  /batches/{id}:
    get:
      operationId: getBatch
      tags: [institutes]
      summary: Get batch by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Batch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Batch"

    patch:
      operationId: updateBatch
      tags: [institutes]
      summary: Update batch
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchUpdate"
      responses:
        "200":
          description: Updated batch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Batch"

    delete:
      operationId: deleteBatch
      tags: [institutes]
      summary: Delete batch
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── PASSAGES ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    User:
      type: object
      required: [id, email, name, role, isActive, subscriptionPlan, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [student, teacher, institute_admin, super_admin]
        phone:
          type: ["string", "null"]
        isActive:
          type: boolean
        subscriptionPlan:
          type: string
          enum: [free, pro_student, institute]
        accountStatus:
          type: string
          enum: [inactive, active, suspended]
        premiumGrantedByOwner:
          type: boolean
        hasAccess:
          type: boolean
          description: Computed server-side — own accountStatus active, OR (if institute member) institute accountStatus active, OR role is super_admin.
        emailVerified:
          type: boolean
        instituteId:
          type: ["integer", "null"]
        instituteName:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...
    UserListResponse:
      type: object
      required: [users, total, page, limit]
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # ─── INSTITUTE SCHEMAS ────────────────────────────────────────────────────
    Institute:
      type: object
      required: [id, name, isActive, subscriptionPlan, accountStatus, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        address:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
        subscriptionPlan:
          type: string
          enum: [free, institute]
        accountStatus:
          type: string
          enum: [inactive, active, suspended]
        premiumGrantedByOwner:
          type: boolean
        referralCode:
          type: ["string", "null"]
        studentCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    InstituteInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 2
        address:
          type: string
        phone:
          type: string
        email:
          type: string
        logoUrl:
          type: string
    InstituteUpdate:
      type: object
      properties:
        name:
          type: string
        address:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
    InstituteListResponse:
      type: object
      required: [institutes, total, page, limit]
      properties:
        institutes:
          type: array
          items:
            $ref: "#/components/schemas/Institute"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    InstituteStats:
      type: object
      required: [totalStudents, totalTests, totalCertificates, avgAccuracy, avgWpm]
      properties:
        totalStudents:
          type: integer
        totalTests:
          type: integer
        totalCertificates:
          type: integer
        avgAccuracy:
          type: number
        avgWpm:
          type: number
        passRate:
          type: number
        recentActivity:
          type: array
          items:
            $ref: "#/components/schemas/ActivityItem"
    ActivityItem:
      type: object
      required: [type, description, createdAt]
      properties:
        type:
          type: string
        description:
          type: string
        userId:
          type: ["integer", "null"]
        userName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time

    # ─── BATCH SCHEMAS ────────────────────────────────────────────────────────
    Batch:
      type: object
      required: [id, name, instituteId, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        instituteId:
          type: integer
        description:
          type: ["string", "null"]
        studentCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    BatchInput:
      type: object
      required: [name, instituteId]
      properties:
        name:
          type: string
          minLength: 2
        instituteId:
          type: integer
        description:
          type: string
    BatchUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]

    # ─── PASSAGE SCHEMAS ──────────────────────────────────────────────────────
...
    AddInstituteStudentInput:
      type: object
      required: [email, password, name]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        name:
          type: string
          minLength: 2
        phone:
          type: string

    # ─── PAYMENT SCHEMAS ──────────────────────────────────────────────────────
...
    SetPremiumInput:
      type: object
      required: [premium]
      properties:
        premium:
          type: boolean
          description: true grants full access immediately, ignoring payment/subscription entirely; false revokes it.
...